gssd: add support for an "allowed-enctypes" option in nfs.conf
Newer kernels have support for newer krb5 encryption types, AES with
SHA2 and Camellia.  An NFS client with an "old" kernel can talk to
and NFS server with a "new" kernel and it just works.   An NFS client
with a "new" kernel can talk to an NFS server with an "old" kernel, but
that requires some additional configuration (particularly if the NFS
server does have support for the newer encryption types in its userspace
krb5 libraries) that may be unclear and/or burdensome to the admin.

1) If the NFS server has support for the newer encryption types in the
   userspace krb5 libraries, but not in the kernel's RPCSEC_GSS code,
   then its possible that it also already has "nfs" keys using those
   newer encryption types in its keytab.  In that case, it's necessary
   to regenerate the "nfs" keys without the newer encryption types.
   The reason this is necessary is because if the NFS client requests
   an "nfs" service ticket from the KDC, and the list of enctypes in
   in that TGS-REQ contains a newer encryption type, and the KDC had
   previously generated a key for the NFS server using the newer
   encryption type, then the resulting service ticket in the TGS-REP
   will be using the newer encryption type and the NFS server will not
   be able to decrypt it.

2) It is necessary to either modify the permitted_enctypes field of the
   krb5.conf or create a custom crypto-policy module (if the
   crypto-policies package is being used) on the NFS *client* so that it
   does not include the newer encryption types.  The reason this is
   necessary is because it affects the list of encryption types that
   will be present in the RPCSEC_GSS_INIT request that the NFS client
   sends to the NFS server.  The kernel on the NFS server cannot not
   process the request on its own; it has to upcall to gssproxy to do
   that... and again if the userspace krb5 libraries on the NFS server
   have support for the newer encryption types, then it will select one
   of those and the kernel will not be able to import the context when
   it gets the downcall.  Also note that modifying the permitted_enctypes
   field and/or crypto policy has the side effect of impacting everything
   krb5 related, not just just NFS.

So add support for an "allowed-enctypes" field in nfs.conf.  This allows
the admin to restrict gssd to using a subset of the encryption types
that are supported by the kernel and krb5 libraries.  This will remove
the need for steps 1 & 2 above, and will only affect NFS rather than
krb5 as a whole.

For example, for a "new" NFS client talking to an "old" NFS server, the
admin will probably want this in the client's nfs.conf:

allowed-enctypes=aes256-cts-hmac-sha1-96,aes128-cts-hmac-sha1-96

Signed-off-by: Scott Mayhew <smayhew@redhat.com>